完善脚本。

+++ build_all.bat
@@ -32,17 +32,17 @@ if exist "%WORK_FOLDER%\tomcat\webapps\ts" rd "%WORK_FOLDER%\tomcat\webapps\ts"
 for /f %%l in ('dir /s /b "%WORK_FOLDER%tsconfig.json"') DO (
   SET FOLDER=%%~dpl
   echo --开始编译：!FOLDER!
+REM 清理js文件
   if not exist !FOLDER!keeps.txt (
-REM echo ---清理目录：!FOLDER!
-    del /s !FOLDER!*.js >nul
+    if exist !FOLDER!*.js del /s !FOLDER!*.js >nul
   )
 REM 不存在keeps.txt文件时，清理已存在js
   for /f %%m in ('dir /b /s /a:d "!FOLDER!"') DO (
     SET FOLDER_SUB=!FOLDER!%%m\
-    echo !FOLDER_SUB!
+REM 清理js文件
     if not exist !FOLDER_SUB!keeps.txt (
-REM  echo ---清理目录：!FOLDER_SUB!
-      del /s !FOLDER_SUB!*.js >nul
+    echo !FOLDER_SUB!keeps.txt
+      if exist !FOLDER_SUB!*.js del /s !FOLDER_SUB!*.js >nul
     )
   )
 REM 运行编译命令
